+++ docker/prod.node-java.dockerfile
